+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.2.0
 
 - `getWordElements()`: fixed a bug when text of a single line with different formatting not handled correctly. **BREAKING CHANGES:** now it will return array of arrays where `[]` - represents line, `[][]` - represents all word nodes of that line. [#4](https://github.com/Amaimersion/google-docs-utils/issues/4)
+- `getSelection()`: fixed a bug when text of a single line with different formatting not handled correctly. **BREAKING CHANGES:** now it will return array of arrays. See documentation for more.
 - `getCaret()`: **BREAKING CHANGES:** `positionIndex` renamed to `positionIndexRelativeToWord`. Now it will represent index relative to word node, not entire line.
 - `getCaretWord()`: fixed a bug when this method not worked with other languages but English. **WARNING:** it still not work with languages which doesn't have upper and lower symbols (Chinese, Japanese, Arabic, Hebrew, etc.).
 - `getLinesTextElements()`, `getLinesText()`, `getCaret()`: fixed a bug when text of a single line with different formatting not handled correctly.